+++ BIDS_converter/config.json
@@ -20,12 +20,11 @@
             , "binary?" : false
             , "binaryEncoding": "float32"
             , "headerData": {"experiment.mat": "channels.name",
-                            "default": "~/Box/CoganLab/ECoG_Task_Data/Timestamps [MASTER].xlsx"
+                            "default": "~/Box/CoganLab/ECoG_Task_Data/Timestamps (MASTER).xlsx"
   }
             , "channels": {"experiment.mat":  "channels"}
             , "sampleRate": "recording.sample_rate"
             , "digital": true
-            , "type": "SEEG"
             , "units": "uV"
             }
 , "task": {"left" : "_?"
@@ -95,7 +94,7 @@
     , "right" : "\\."
     , "content" : ["[0-9]{2}"]
     }
-, "acq" : {"left" : "((Session)|(Part))"
+, "acq" : {"left" : "((Session)|(Part)|(part))"
     , "right" : "[_ \\.]"
     , "content" : ["[0-9]{1,3}"]
     , "fill": 2
